airflow/requirements-dashboard.txt
dash
plotly
pandas
psycopg2-binary
gradio